chat_flask_fornt_backend/backend/requirements.txt
aiohttp==3.9.5
aiohttp-socks==0.8.4
aiosignal==1.3.1
alembic==1.13.1
annotated-types==0.7.0
anyio==4.4.0
api==0.0.7
async-timeout==4.0.3
attrs==23.2.0
bcrypt==4.1.2
beautifulsoup4==4.12.3
bidict==0.23.1
blinker==1.8.1
bottle==0.12.25
Brotli==1.1.0
browser-cookie3==0.19.1
cairocffi==1.7.1
CairoSVG==2.7.1
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
clr-loader==0.2.6
colorama==0.4.6
cryptography==43.0.0
cssselect2==0.7.0
curl_cffi==0.7.1
defusedxml==0.7.1
distro==1.9.0
dnspython==2.6.1
duckduckgo_search==6.2.4
email_validator==2.2.0
fastapi==0.111.1
fastapi-cli==0.0.4
Flask==3.0.3
Flask-API==3.1
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.0
Flask-JWT-Extended==4.6.0
flask-marshmallow==1.2.1
Flask-Migrate==4.0.7
Flask-PyMongo==2.3.0
Flask-SocketIO==5.3.6
Flask-SQLAlchemy==3.1.1
frozenlist==1.4.1
greenlet==3.0.3
gunicorn==22.0.0
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.3
loguru==0.7.2
lz4==4.3.3
Mako==1.3.3
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.1
marshmallow-sqlalchemy==1.0.0
mdurl==0.1.2
multidict==6.0.5
nest-asyncio==1.6.0
nose==1.3.7
OpenAIAuth==3.0.0
outcome==1.3.0.post0
pillow==10.4.0
platformdirs==4.2.2
plyer==2.1.0
primp==0.5.4
prompt_toolkit==3.0.47
proxy_tools==0.1.0
psycopg2==2.9.9
pycparser==2.22
pycryptodome==3.20.0
pycryptodomex==3.20.0
pydantic==2.8.2
pydantic_core==2.20.1
PyExecJS==1.5.1
Pygments==2.18.0
PyJWT==2.8.0
pymongo==4.8.0
PySocks==1.7.1
python-dotenv==1.0.1
python-engineio==4.9.0
python-multipart==0.0.9
python-socketio==5.11.2
python-socks==2.5.0
pythonnet==3.0.3
pywebview==5.1
PyYAML==6.0.1
regex==2024.7.24
requests==2.32.3
rich==13.7.1
shellingham==1.5.4
simple-websocket==1.0.0
six==1.16.0
sniffio==1.3.1
socksio==1.0.0
sortedcontainers==2.4.0
soupsieve==2.5
SQLAlchemy==2.0.29
starlette==0.37.2
tiktoken==0.7.0
tinycss2==1.3.0
tls-client==1.0.1
tqdm==4.66.4
trio==0.26.0
trio-websocket==0.11.1
typer==0.12.3
typing_extensions==4.12.2
undetected-chromedriver==3.5.5
undetected-edgedriver==3.1.6.4
urllib3==2.2.2
uvicorn==0.30.3
waitress==3.0.0
watchfiles==0.22.0
wcwidth==0.2.13
webencodings==0.5.1
websocket-client==1.8.0
websockets==12.0
Werkzeug==3.0.3
win32-setctime==1.1.0
wsproto==1.2.0
yarl==1.9.4
